test(layout): extract large inline test modules
